# docs/swagger.yaml
definitions:
...
  fitness-tracker-backend_workout_handler.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.replaceWorkoutSetsRequest:
    properties:
      sets:
        items:
          $ref: '#/definitions/handler.workoutSetRequest'
        type: array
    type: object
...
  handler.workoutSessionRequest:
    properties:
      datetime:
        type: string
      order_index:
        type: integer
      sets:
        items:
          $ref: '#/definitions/handler.workoutSetRequest'
        type: array
      training_id:
        type: integer
      workout_type_id:
        type: integer
    required:
    - training_id
    - workout_type_id
    type: object
  handler.workoutSetRequest:
    properties:
      distance_meters:
        type: number
      duration_seconds:
        type: integer
      notes:
        type: string
      reps:
        type: integer
      set_number:
        type: integer
      weight_kg:
        type: number
    type: object
...
  models.MuscleGroup:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Training:
    properties:
      created_at:
        type: string
      finished_at:
        type: string
      id:
        type: integer
      notes:
        type: string
      sessions:
        items:
          $ref: '#/definitions/models.WorkoutSession'
        type: array
      started_at:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  models.WorkoutSession:
    properties:
      datetime:
        type: string
      id:
        type: integer
      order_index:
        type: integer
      sets:
        items:
          $ref: '#/definitions/models.WorkoutSet'
        type: array
      training:
        allOf:
        - $ref: '#/definitions/models.Training'
        description: Associations
      training_id:
        type: integer
      user_id:
        type: integer
      workout_type:
        $ref: '#/definitions/models.WorkoutType'
      workout_type_id:
        type: integer
    type: object
  models.WorkoutSet:
    properties:
      distance_meters:
        type: number
      duration_seconds:
        type: integer
      id:
        type: integer
      notes:
        type: string
      reps:
        type: integer
      set_number:
        type: integer
      weight_kg:
        type: number
      workout_session_id:
        type: integer
    type: object
  models.WorkoutType:
    properties:
      default_metric:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      muscle_group:
        allOf:
        - $ref: '#/definitions/models.MuscleGroup'
        description: Associations
      muscle_group_id:
        type: integer
      name:
        type: string
    type: object
info:
  contact: {}
  description: API documentation for the Fitness Tracker backend service.
  title: Fitness Tracker API
  version: "1.0"
paths:
...
  /workout-sessions:
    get:
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Training ID
        in: query
        name: training_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.WorkoutSession'
            type: array
      security:
      - BearerAuth: []
      summary: List workout sessions for user
      tags:
      - workout-sessions
    post:
      consumes:
      - application/json
      parameters:
      - description: Session
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.workoutSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.WorkoutSession'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Create workout session
      tags:
      - workout-sessions
  /workout-sessions/{id}:
    delete:
      parameters:
      - description: WorkoutSession ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete workout session
      tags:
      - workout-sessions
    get:
      parameters:
      - description: WorkoutSession ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.WorkoutSession'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Get workout session by ID
      tags:
      - workout-sessions
  /workout-sessions/{id}/sets:
    get:
      parameters:
      - description: WorkoutSession ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.WorkoutSet'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: List sets for a workout session
      tags:
      - workout-sessions
    put:
      consumes:
      - application/json
      parameters:
      - description: WorkoutSession ID
        in: path
        name: id
        required: true
        type: integer
      - description: Sets
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.replaceWorkoutSetsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.WorkoutSession'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Replace all sets for a workout session
      tags:
      - workout-sessions
...
securityDefinitions:
  BearerAuth:
    description: 'Provide your JWT with the "Bearer " prefix. Example: "Bearer {token}".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"